refactor: change path datagrid
