+++ src/components/dashboard/admin/client/form/list/index.js
@@ -5,11 +5,10 @@ import DialogActions from '@material-ui/core/DialogActions'
 import DialogContent from '@material-ui/core/DialogContent'
 import DialogTitle from '@material-ui/core/DialogTitle'
 
-import DataList from '~/components/datagrid'
+import DataList from '~/components/dashboard/datagrid'
 
 const ListAddress = ({ open, close, data }) => {
   const columnAddress = [
-   
     {
       field: 'city',
       headerName: 'Cidade',
