+++ src/components/dashboard/admin/solicitations/index.js
@@ -3,7 +3,7 @@ import { IconButton, Tooltip } from '@material-ui/core'
 import { DataGrid, GridToolbar } from '@mui/x-data-grid'
 import { More as MoreIcon } from '@mui/icons-material'
 import Loading from '../../../loading/page/index'
-import { BoxTable } from '../../datagrid/styled'
+import { BoxTable } from '../../datalist/styled'
 import ListProduct from './details/index'
 import ListDelivery from './delivery/index'
