+++ src/components/dashboard/datagrid/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import { DataGrid } from '@material-ui/data-grid'
-import { BoxTable } from './styled'
+import { DataGrid, GridToolbar } from '@mui/x-data-grid'
 import Loading from '../../loading/page/index'
+import { BoxTable } from './styled'
 
 const DataList = ({ data, columns, loading }) => {
   if (loading) {
@@ -10,7 +10,21 @@ const DataList = ({ data, columns, loading }) => {
 
   return (
     <BoxTable>
-      <DataGrid rows={data} columns={columns} pageSize={10} />
+      <DataGrid
+        rows={data}
+        columns={columns}
+        loading={loading}
+        pageSize={10}
+        disableColumnSelector
+        disableDensitySelector
+        components={{ Toolbar: GridToolbar }}
+        componentsProps={{
+          toolbar: {
+            showQuickFilter: true,
+            quickFilterProps: { debounceMs: 500 }
+          }
+        }}
+      />
     </BoxTable>
   )
 }
