+++ src/components/dashboard/datalist/styled.js
@@ -4,3 +4,7 @@ export const BoxTable = styled.div`
   height: 600px;
   width: 100%;
 `
+export const SImg = styled.img`
+  width: 60px;
+  height: 45px;
+`
